changed exception messages to their respective classes and minor refactoring

+++ src/main/java/com/github/allanccruz/POC1/api/service/CustomerService.java
@@ -1,8 +1,8 @@
 package com.github.allanccruz.POC1.api.service;
 
-import com.github.allanccruz.POC1.api.Exception.CustomerNotFoundException;
 import com.github.allanccruz.POC1.api.dto.request.CustomerRequestDto;
 import com.github.allanccruz.POC1.api.entities.Customer;
+import com.github.allanccruz.POC1.api.exception.customerException.CustomerNotFoundException;
 import com.github.allanccruz.POC1.api.repository.CustomerRepository;
 import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -27,7 +26,7 @@ public class CustomerService {
     }
 
     public Customer findById(UUID id) {
-        return customerRepository.findById(id).orElseThrow(() -> new CustomerNotFoundException("Id not found"));
+        return customerRepository.findById(id).orElseThrow(() -> new CustomerNotFoundException(id));
     }
 
     public List<Customer> findAll() {
@@ -36,8 +35,9 @@ public class CustomerService {
 
     @Transactional
     public void deleteById(UUID id) {
-        Optional<Customer> customerOptional = customerRepository.findById(id);
-        customerOptional.orElseThrow(() -> new CustomerNotFoundException("Customer does not exist!"));
+        if (customerRepository.findById(id).isEmpty()) {
+            throw new CustomerNotFoundException(id);
+        }
         customerRepository.deleteById(id);
     }
 
@@ -46,7 +46,7 @@ public class CustomerService {
         if (existById(customerRequestDto.getId())) {
             return customerRepository.save(mapper.map(customerRequestDto, Customer.class));
         } else {
-            throw new CustomerNotFoundException("Customer does not exist!");
+            throw new CustomerNotFoundException(customerRequestDto.getId());
         }
     }
